% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addXLreadme.R
\name{.addXLreadme}
\alias{.addXLreadme}
\title{Add a readme sheet to the Excel workbook}
\usage{
.addXLreadme(wbList = XL.wb)
}
\arguments{
\item{wbList}{list with workbook and default cell styles (i.e., output from getXLsettings)}
}
\value{
list with workbook and default cell styles
}
\description{
Add a readme sheet to the Excel workbook
}
